check-mod-tidy: use git worktree instead of copy
Hopefully this is more reliable than a direct copy. We have multiple
processes working concurrently in the checkout and it sometimes results
in failures like this one:
https://build.dev.storj.io/blue/organizations/jenkins/storj-gerrit/detail/storj-gerrit/6088/pipeline/

+++ check-mod-tidy/main.go
@@ -26,16 +26,25 @@ func check() (err error) {
 	}
 
 	defer func() {
-		err := os.RemoveAll(tempDir)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "failed to delete temporary directory: %v\n", err)
+		if _, err := os.Stat(tempDir); err == nil {
+			err := os.RemoveAll(tempDir)
+			if err != nil {
+				fmt.Fprintf(os.Stderr, "failed to delete temporary directory: %v\n", err)
+			}
 		}
 	}()
 
-	err = copyDir(".", tempDir)
+	err = worktreeAdd(tempDir)
 	if err != nil {
-		return fmt.Errorf("failed to copy dir: %w", err)
+		return fmt.Errorf("failed to add worktree: %w", err)
 	}
+
+	defer func() {
+		err = worktreeRemove(tempDir)
+		if err != nil {
+			fmt.Fprintf(os.Stderr, "failed to remove worktree: %v\n", err)
+		}
+	}()
 
 	err = os.Chdir(tempDir)
 	if err != nil {
@@ -84,8 +93,15 @@ func tidy() (err error) {
 	return err
 }
 
-func copyDir(src, dst string) error {
-	cmd := exec.Command("cp", "-a", src, dst)
+func worktreeAdd(dst string) (err error) {
+	cmd := exec.Command("git", "worktree", "add", "--detach", dst)
+	cmd.Stdout, cmd.Stderr = os.Stderr, os.Stderr
+
+	return cmd.Run()
+}
+
+func worktreeRemove(dst string) (err error) {
+	cmd := exec.Command("git", "worktree", "remove", "--force", dst)
 	cmd.Stdout, cmd.Stderr = os.Stderr, os.Stderr
 
 	return cmd.Run()
